Refactor: ScopeSelector eventHandler분리
