+++ src/layout/Header/Section/SearchBox/ScopeSelector/index.js
@@ -1,9 +1,6 @@
 import HtmlElement from '../../../../../utils/HtmlElement.js';
-import {
-  findTargetIdElement,
-  handleDisplayElement,
-} from '../../../../../utils/manuplateDOM.js';
 import { POP_UP } from '../../../../../constant.js';
+import eventHandler from './eventHandler.js';
 
 export default function ScopeSelector($element, args) {
   HtmlElement.call(this, $element, args);
@@ -12,26 +9,22 @@ export default function ScopeSelector($element, args) {
 ScopeSelector.prototype = Object.create(HtmlElement.prototype);
 ScopeSelector.prototype.constructor = ScopeSelector;
 
+ScopeSelector.prototype.init = function () {
+  this.state = {
+    ...this.args,
+  };
+  this.eventHandler = eventHandler;
+};
+
 ScopeSelector.prototype.setTemplate = function () {
   const { option } = this.state;
   return template(option);
 };
 
 ScopeSelector.prototype.setEvent = function () {
+  const { handleClick } = this.eventHandler;
   this.$element.addEventListener('click', handleClick.bind(this));
 };
-
-function handleClick({ target }) {
-  const $searchSelector = target.closest('#searchSelector');
-  if ($searchSelector) return showCategory.apply(this);
-  if (target.dataset?.option)
-    return this.state.changeSearchOption(target.dataset.option);
-}
-
-function showCategory() {
-  const $options = findTargetIdElement(this.$element, 'searchOptions');
-  handleDisplayElement($options);
-}
 
 const options = [
   '전체',
